+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -487,7 +487,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KY2MFWKTPFCWH96M487C9EQA"),
+  _id: z.string().default("01KY3S9PGDYK9WT8192PCH8CPK"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: CreateDatasourceStatus$inboundSchema,